rename: model => domain (package 리네임)
