+++ src/main/java/db/DataBase.java
@@ -1,7 +1,7 @@
 package db;
 
 import com.google.common.collect.Maps;
-import model.User;
+import domain.User;
 
 import java.util.Collection;
 import java.util.Map;
