+++ src/main/java/domain/User.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.io.Serializable;
