+++ src/test/java/servlet/controller/UserLoginControllerTest.java
@@ -3,8 +3,8 @@ package servlet.controller;
 import http.request.HttpRequest;
 import http.request.HttpRequestFactory;
 import http.response.HttpResponse;
-import model.User;
-import model.UserService;
+import domain.User;
+import domain.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
